+++ front/README.md
@@ -1,0 +1,19 @@
+# Front-End code for commit.gg (temp)
+
+## You need
+
+- node v10.15.3
+- yarn v1.16.0
+
+## Development
+
+```sh
+yarn run start
+```
+
+Can access via port :8080
+
+## Production
+```sh
+yarn run build
+```
