+++ front/package.json
@@ -18,11 +18,13 @@
     "@hot-loader/react-dom": "^16.8.6",
     "@types/react": "^16.8.23",
     "@types/react-dom": "^16.8.4",
+    "@types/react-router-dom": "^4.3.4",
     "@types/styled-components": "^4.1.18",
     "core-js": "^3.1.4",
     "normalize.css": "^8.0.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.1",
     "regenerator-runtime": "^0.13.2",
     "styled-components": "^4.3.2",
     "three": "^0.106.2"
@@ -51,7 +53,7 @@
     "html-webpack-plugin": "^3.2.0",
     "husky": ">=1",
     "lint-staged": ">=8",
-    "mini-css-extract-plugin": "^0.7.0",
+    "mini-css-extract-plugin": "^0.8.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "prettier": "^1.18.2",
     "react-hot-loader": "^4.12.3",
