+++ front/src/index.tsx
@@ -4,7 +4,13 @@ import 'react-hot-loader';
 import 'normalize.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
 
 import App from './pages/App';
 
-ReactDOM.render(<App />, document.getElementById('root') as HTMLElement);
+ReactDOM.render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>,
+  document.getElementById('root') as HTMLElement
+);
