+++ front/src/pages/User.tsx
@@ -1,0 +1,8 @@
+import React, { FC } from 'react';
+import { RouteComponentProps } from 'react-router';
+
+const User: FC<RouteComponentProps> = ({ match }) => {
+  return <div>wow</div>;
+};
+
+export default User;
